# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/ChangSZ/mall-go/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/menu/sort": {
            "patch": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "更新菜单排序",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.menu"
                ],
                "summary": "更新菜单排序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hashId",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "排序",
                        "name": "sort",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.updateSortResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
        "/api/menu/{id}": {
            "get": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "菜单详情",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.menu"
                ],
                "summary": "菜单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hashId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.detailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "删除菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.menu"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hashId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.deleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
        "/api/menu_action": {
            "get": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "功能权限列表",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.menu"
                ],
                "summary": "功能权限列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hashID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.listActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "创建功能权限",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.menu"
                ],
                "summary": "创建功能权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "HashID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "请求方法",
                        "name": "method",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "请求地址",
                        "name": "api",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menu.createActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
        "/menu/{id}": {
            "get": {
                "description": "根据ID获取菜单详情",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UmsMenuController"
                ],
                "summary": "根据ID获取菜单详情",
                "parameters": [
                    {
                        "description": "请求信息",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_menu.getRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ums_menu.getResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "code.Failure": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "业务返回数据"
                },
                "message": {
                    "description": "描述信息",
                    "type": "string"
                }
            }
        },
        "code.Success": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "业务返回数据"
                },
                "message": {
                    "description": "描述信息",
                    "type": "string"
                }
            }
        },
...
        "menu.createActionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
...
        "menu.deleteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
        "menu.detailResponse": {
            "type": "object",
            "properties": {
                "icon": {
                    "description": "图标",
                    "type": "string"
                },
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "link": {
                    "description": "链接地址",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "pid": {
                    "description": "父类ID",
                    "type": "integer"
                }
            }
        },
        "menu.listActionData": {
            "type": "object",
            "properties": {
                "api": {
                    "description": "调用方对接人",
                    "type": "string"
                },
                "hash_id": {
                    "description": "hashID",
                    "type": "string"
                },
                "menu_id": {
                    "description": "菜单栏ID",
                    "type": "integer"
                },
                "method": {
                    "description": "调用方secret",
                    "type": "string"
                }
            }
        },
        "menu.listActionResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/menu.listActionData"
                    }
                },
                "menu_name": {
                    "type": "string"
                }
            }
        },
...
        "menu.updateSortResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
...
        "ums_menu.getRequest": {
            "type": "object"
        },
        "ums_menu.getResponse": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "LoginToken": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}